requirements.txt
botocore==1.12.25
cached-property==1.3.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
coverage==3.7.1
cryptography==1.7.2
enum34==1.1.6
funcsigs==1.0.2
ipaddress==1.0.19
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
pyasn1==0.4.2
PyYAML==3.12
redis==3.2.0
requests==2.21.0
simplejson==3.13.2
urllib3==1.24.1
websocket-client==0.44.0